+++ src/Contracts/LocalizedApplication.php
@@ -5,7 +5,7 @@ namespace Goodcat\L10n\Contracts;
 interface LocalizedApplication
 {
     /** @see \Goodcat\L10n\Mixin\LocalizedApplication::getPreferredLocale */
-    public function getPreferredLocale(): string;
+    public function getPreferredLocale(): ?string;
 
     /** @see \Goodcat\L10n\Mixin\LocalizedApplication::setPreferredLocale */
     public function setPreferredLocale(string $locale): void;